list.py: Renamed list module to lists to avoid conflict with builtin.
