+++ src/ponysay.py
@@ -35,7 +35,7 @@ from balloon import *
 from spellocorrecter import *
 from ucs import *
 from kms import *
-import list as _list
+import lists
 from metadata import *
 
 
@@ -739,7 +739,7 @@ class Ponysay():
         
         @param  ponydirs:itr<str>?  The pony directories to use
         '''
-        _list.simplelist(self.ponydirs if ponydirs is None else ponydirs,
+        lists.simplelist(self.ponydirs if ponydirs is None else ponydirs,
                         self.__quoters(), lambda x : self.__ucsise(x))
     
     
@@ -749,7 +749,7 @@ class Ponysay():
         
         @param  ponydirs:itr<str>  The pony directories to use
         '''
-        _list.linklist(self.ponydirs if ponydirs is None else ponydirs,
+        lists.linklist(self.ponydirs if ponydirs is None else ponydirs,
                       self.__quoters(), lambda x, y : self.__ucsise(x, y))
     
     
@@ -760,7 +760,7 @@ class Ponysay():
         @param  standard:bool  Include standard ponies
         @param  extra:bool     Include extra ponies
         '''
-        _list.onelist(self.ponydirs if standard else None,
+        lists.onelist(self.ponydirs if standard else None,
                      self.extraponydirs if extra else None,
                      lambda x : self.__ucsise(x))
     
@@ -801,7 +801,7 @@ class Ponysay():
         '''
         Prints a list of all balloons
         '''
-        _list.balloonlist(self.balloondirs, self.isthink)
+        lists.balloonlist(self.balloondirs, self.isthink)
     
     
     def __getBalloonPath(self, names, alt = False):
